Remove unused imports from several Java classes

+++ plugins/org.locationtech.udig.printing.ui/src/org/locationtech/udig/printing/ui/internal/PdfPrintingEngine.java
@@ -1,7 +1,7 @@
-/*
- *    uDig - User Friendly Desktop Internet GIS client
- *    http://udig.refractions.net
- *    (C) 2004, Refractions Research Inc.
+/**
+ * uDig - User Friendly Desktop Internet GIS client
+ * http://udig.refractions.net
+ * (C) 2004, Refractions Research Inc.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -11,24 +11,18 @@
  */
 package org.locationtech.udig.printing.ui.internal;
 
-import java.awt.Graphics;
 import java.awt.Graphics2D;
-import java.awt.geom.AffineTransform;
-import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
 
-import javax.imageio.ImageIO;
-
-import org.locationtech.udig.printing.model.Box;
-import org.locationtech.udig.printing.model.BoxPrinter;
-import org.locationtech.udig.printing.model.Page;
-
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.draw2d.geometry.Dimension;
 import org.eclipse.draw2d.geometry.Point;
+import org.locationtech.udig.printing.model.Box;
+import org.locationtech.udig.printing.model.BoxPrinter;
+import org.locationtech.udig.printing.model.Page;
 
 import com.lowagie.text.Document;
 import com.lowagie.text.DocumentException;
@@ -37,23 +31,25 @@ import com.lowagie.text.pdf.PdfContentByte;
 import com.lowagie.text.pdf.PdfWriter;
 
 /**
- * The engine that prints to pdf.
- * 
+ * The engine that prints to PDF.
+ *
  * @author Andrea Antonello (www.hydrologis.com)
  */
 public class PdfPrintingEngine {
 
     private Page page;
+
     private IProgressMonitor monitor;
+
     private final File outputPdfFile;
 
     /**
      * Constructs a PdfPrintingEngine using the given Page and the file to which to dump to.
-     * 
+     *
      * @param page the Page to be printed.
      * @param outputPdfFile the file to which to dump to.
      */
-    public PdfPrintingEngine( Page page, File outputPdfFile ) {
+    public PdfPrintingEngine(Page page, File outputPdfFile) {
         this.page = page;
         this.outputPdfFile = outputPdfFile;
     }
@@ -61,7 +57,7 @@ public class PdfPrintingEngine {
     /**
      * @param monitor
      */
-    public void setMonitor( IProgressMonitor monitor ) {
+    public void setMonitor(IProgressMonitor monitor) {
         this.monitor = monitor;
     }
 
@@ -82,31 +78,28 @@ public class PdfPrintingEngine {
             document.open();
 
             PdfContentByte cb = writer.getDirectContent();
-            Graphics2D graphics = cb
-                    .createGraphics(paperRectangle.getWidth(), paperRectangle.getHeight());
-
-            // BufferedImage bI = new BufferedImage((int) paperRectangle.width(), (int)
-            // paperRectangle
-            // .height(), BufferedImage.TYPE_INT_ARGB);
-            // Graphics graphics2 = bI.getGraphics();
+            Graphics2D graphics = cb.createGraphics(paperRectangle.getWidth(),
+                    paperRectangle.getHeight());
 
             List<Box> boxes = page.getBoxes();
-            for( Box box : boxes ) {
+            for (Box box : boxes) {
                 String id = box.getID();
                 System.out.println(id);
                 Point boxLocation = box.getLocation();
-                if( boxLocation == null ) continue;
+                if (boxLocation == null)
+                    continue;
                 int x = boxLocation.x;
                 int y = boxLocation.y;
                 Dimension size = box.getSize();
-                if( size == null ) continue; 
+                if (size == null)
+                    continue;
                 int w = size.width;
                 int h = size.height;
 
-                float newX = xScale * (float) x;
-                float newY = yScale * (float) y;
-                float newW = xScale * (float) w;
-                float newH = yScale * (float) h;
+                float newX = xScale * x;
+                float newY = yScale * y;
+                float newW = xScale * w;
+                float newH = yScale * h;
 
                 box.setSize(new Dimension((int) newW, (int) newH));
                 box.setLocation(new Point((int) newX, (int) newY));
@@ -118,8 +111,6 @@ public class PdfPrintingEngine {
             }
 
             graphics.dispose();
-            // ImageIO.write(bI, "png", new File("c:\\Users\\moovida\\Desktop\\test.png"));
-            // graphics.drawImage(bI, null, 0, 0);
 
             document.newPage();
             document.close();
